Move utils to reduce code duplication
