+++ internal/cluster/sbd.go
@@ -12,6 +12,7 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/pkg/errors"
+	"github.com/trento-project/agent/internal/utils"
 )
 
 const (
@@ -55,38 +56,6 @@ type SBDNode struct {
 
 var sbdDumpExecCommand = exec.Command //nolint
 var sbdListExecCommand = exec.Command //nolint
-
-// FindMatches finds regular expression matches in a key/value based
-// text (ini files, for example), and returns a map with them.
-// If the matched key has spaces, they will be replaced with underscores
-// If the same keys is found multiple times, the entry of the map will
-// have a list as value with all of the matched values
-// The pattern must have 2 groups. For example: `(.+)=(.*)`
-func FindMatches(pattern string, text []byte) map[string]interface{} {
-	configMap := make(map[string]interface{})
-
-	r := regexp.MustCompile(pattern)
-	values := r.FindAllStringSubmatch(string(text), -1)
-	for _, match := range values {
-		key := strings.ReplaceAll(match[1], " ", "_")
-		if _, ok := configMap[key]; ok {
-			switch configMap[key].(type) {
-			case string:
-				{
-					configMap[key] = []interface{}{configMap[key]}
-				}
-			default:
-				{
-					configMap[key] = append(configMap[key].([]interface{}), match[2]) //nolint please check me for safe casting
-
-				}
-			}
-		} else {
-			configMap[key] = match[2]
-		}
-	}
-	return configMap
-}
 
 func NewSBD(cluster, sbdPath, sbdConfigPath string) (SBD, error) {
 	var s = SBD{
@@ -134,7 +103,7 @@ func getSBDConfig(sbdConfigPath string) (map[string]interface{}, error) {
 		return nil, errors.Wrap(err, "could not read sbd config file")
 	}
 
-	configMap := FindMatches(`(?m)^(\w+)=(\S[^#\s]*)`, sbdConfigRaw)
+	configMap := utils.FindMatches(`(?m)^(\w+)=(\S[^#\s]*)`, sbdConfigRaw)
 
 	return configMap, nil
 }
